*: undo the change to controller_context
EtcdInformer is only used by the MCO pod to reconcile image names.
This pulls out the logic of creating etcd informer from the generic
controller context to MCO pod start method.
